+++ NEWS.md
@@ -1,7 +1,12 @@
+# divraster 1.0.3
+
+* Improvements in function descriptions.
+* Improvements in vignette (examples and description).
+* Minor correction in function tests [#2](https://github.com/flaviomoc/divraster/issues/2).
+* Citation file added.
+
 # divraster 1.0.2
 
-# divraster 1.0.1
-
-# divraster 1.0.0
+Released 2023-06-27
 
 * Initial CRAN submission.